beats-biopods-sentiment-starter/requirements.txt
pandas
numpy
scikit-learn
matplotlib
nltk
spacy
wordcloud
plotly
streamlit
pyyaml
